+++ TaskPlannerMetrum/TaskPlannerMetrum/Repository/Generic/IRepository.cs
@@ -6,10 +6,10 @@ namespace TaskPlannerMetrum.Repository.Generic
     public interface IRepository<T> where T : BaseEntity
     {
         T Create(T item);
-        T FindByID(long id);
+        T FindByID(int id);
         List<T> FindAll();
         T Update(T item);
-        void Delete(long id);
-        bool Exists(long id);
+        void Delete(int id);
+        bool Exists(int id);
     }
 }
